Location-aware room listing.
Now that the possibility exists on the backend, rooms are shown to the user by
increasing order of distance. This means that we need to get the user's location
somehow, etc...

Note that we're not yet sending the location when creating a new room. This will
have to be addressed in a subsequent diff.