Improve playing of tracks after completion.
This feature was already present before, but didn't work on the HTC Desire HD
when the phone was on standby.  Instead of restarting the MainActivity with a
special intent, we use broadcasted intents. This also removes the coupling
between MusicService and MainActivity.

In addition, we stop the music playback when we leave the room (that is, when
PlayerFragment is destroyed).