Changed the way the track queue works.
To reflect the server side changes and the fact that we generate a "static"
playlist, the TrackQueue class had to be significantly modified. Now, we receive
a chunk of the playlist instead of the next track only. We also pay attention to
not play the same track twice, and to play the tracks in the order in which they
were sent by the recommender system.